models/video/cross_efficient_vit/requirements.txt
numpy>=1.23.0,<2.0
Pillow>=8.3.1
scikit-learn
albumentations==0.5.2
tqdm
pyyaml
efficientnet_pytorch==0.7.1
einops
facenet-pytorch==2.5.2
gdown

fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0,<3.0.0
python-multipart
requests
